# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/businesses/me/setup": {
      "get": {
        "operationId": "BusinessesSetupController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Business setup profile, legal profile, and default branch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSetupResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get current business setup profile",
        "tags": [
          "Businesses"
        ]
      },
      "patch": {
        "operationId": "BusinessesSetupController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessSetupBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated business setup profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSetupResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update current business setup profile",
        "tags": [
          "Businesses"
        ]
      }
    },
    "/v1/branches": {
      "get": {
        "operationId": "BranchesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tenant branch list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List branches for the current business",
        "tags": [
          "Branches"
        ]
      },
      "post": {
        "operationId": "BranchesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created branch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a branch",
        "tags": [
          "Branches"
        ]
      }
    },
    "/v1/branches/{branchId}": {
      "get": {
        "operationId": "BranchesController_get",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Branch profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Branch not found."
          }
        },
        "summary": "Get a branch",
        "tags": [
          "Branches"
        ]
      },
      "patch": {
        "operationId": "BranchesController_update",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated branch.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a branch",
        "tags": [
          "Branches"
        ]
      },
      "delete": {
        "operationId": "BranchesController_delete",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Branch deactivated."
          }
        },
        "summary": "Deactivate a branch",
        "tags": [
          "Branches"
        ]
      }
    },
    "/v1/branches/{branchId}/set-default": {
      "post": {
        "operationId": "BranchesController_setDefault",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default branch updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchResponseDto"
                }
              }
            }
          }
        },
        "summary": "Set branch as default",
        "tags": [
          "Branches"
        ]
      }
    },
...
  },
  "info": {
    "title": "Quickarte API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001/v1"
    }
  ],
  "components": {
    "schemas": {
...
      "BusinessSetupBusinessDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "restaurant",
              "cafe",
              "autre"
            ]
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "logo": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string"
          },
          "secondaryCurrency": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "type",
          "currency",
          "timezone",
          "locale"
        ]
      },
      "BusinessLegalProfileDto": {
        "type": "object",
        "properties": {
          "legalName": {
            "type": "string"
          },
          "iceNumber": {
            "type": "string",
            "nullable": true
          },
          "rcNumber": {
            "type": "string",
            "nullable": true
          },
          "ifNumber": {
            "type": "string",
            "nullable": true
          },
          "patenteNumber": {
            "type": "string",
            "nullable": true
          },
          "cnssNumber": {
            "type": "string",
            "nullable": true
          },
          "legalAddress": {
            "type": "string",
            "nullable": true
          },
          "legalCity": {
            "type": "string",
            "nullable": true
          },
          "legalPostcode": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "legalName"
        ]
      },
      "DefaultBranchSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "isDefault"
        ]
      },
      "BusinessSetupResponseDto": {
        "type": "object",
        "properties": {
          "business": {
            "$ref": "#/components/schemas/BusinessSetupBusinessDto"
          },
          "legalProfile": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessLegalProfileDto"
              }
            ]
          },
          "defaultBranch": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DefaultBranchSummaryDto"
              }
            ]
          }
        },
        "required": [
          "business"
        ]
      },
      "LegalProfileBody": {
        "type": "object",
        "properties": {
          "legalName": {
            "type": "string"
          },
          "iceNumber": {
            "type": "string",
            "nullable": true,
            "maxLength": 32
          },
          "rcNumber": {
            "type": "string",
            "nullable": true,
            "maxLength": 32
          },
          "ifNumber": {
            "type": "string",
            "nullable": true,
            "maxLength": 32
          },
          "patenteNumber": {
            "type": "string",
            "nullable": true,
            "maxLength": 32
          },
          "cnssNumber": {
            "type": "string",
            "nullable": true,
            "maxLength": 32
          },
          "legalAddress": {
            "type": "string",
            "nullable": true
          },
          "legalCity": {
            "type": "string",
            "nullable": true
          },
          "legalPostcode": {
            "type": "string",
            "nullable": true,
            "maxLength": 16
          }
        }
      },
      "UpdateBusinessSetupBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "restaurant",
              "cafe",
              "autre"
            ]
          },
          "currency": {
            "type": "string"
          },
          "secondaryCurrency": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "logo": {
            "type": "string",
            "nullable": true
          },
          "legalProfile": {
            "$ref": "#/components/schemas/LegalProfileBody"
          }
        }
      },
      "BranchResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "businessId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "postcode": {
            "type": "string",
            "nullable": true
          },
          "countryCode": {
            "type": "string"
          },
          "googlePlaceId": {
            "type": "string",
            "nullable": true
          },
          "formattedAddress": {
            "type": "string",
            "nullable": true
          },
          "lat": {
            "type": "string",
            "nullable": true
          },
          "lng": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "socialLinks": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "logo": {
            "type": "string",
            "nullable": true
          },
          "cuisineType": {
            "type": "string",
            "nullable": true
          },
          "seatingCapacity": {
            "type": "number",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "nullable": true
          },
          "locale": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "businessId",
          "name",
          "slug",
          "isDefault",
          "status",
          "countryCode",
          "createdAt",
          "updatedAt"
        ]
      },
      "BranchListResponseDto": {
        "type": "object",
        "properties": {
          "branches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchResponseDto"
            }
          }
        },
        "required": [
          "branches"
        ]
      },
      "BranchBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "slug": {
            "type": "string",
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "postcode": {
            "type": "string",
            "nullable": true,
            "maxLength": 16
          },
          "countryCode": {
            "type": "string",
            "maxLength": 2,
            "default": "MA"
          },
          "googlePlaceId": {
            "type": "string",
            "nullable": true
          },
          "formattedAddress": {
            "type": "string",
            "nullable": true
          },
          "lat": {
            "type": "string",
            "nullable": true
          },
          "lng": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true,
            "format": "email"
          },
          "website": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "socialLinks": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "logo": {
            "type": "string",
            "nullable": true
          },
          "cuisineType": {
            "type": "string",
            "nullable": true
          },
          "seatingCapacity": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "nullable": true
          },
          "locale": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateBranchBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "slug": {
            "type": "string",
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "postcode": {
            "type": "string",
            "nullable": true,
            "maxLength": 16
          },
          "countryCode": {
            "type": "string",
            "maxLength": 2
          },
          "googlePlaceId": {
            "type": "string",
            "nullable": true
          },
          "formattedAddress": {
            "type": "string",
            "nullable": true
          },
          "lat": {
            "type": "string",
            "nullable": true
          },
          "lng": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true,
            "format": "email"
          },
          "website": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          },
          "socialLinks": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "logo": {
            "type": "string",
            "nullable": true
          },
          "cuisineType": {
            "type": "string",
            "nullable": true
          },
          "seatingCapacity": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "nullable": true
          },
          "locale": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}